move test case for welcome page to test index file
