+++ tests/app/main/views/test_index.py
@@ -295,3 +295,14 @@ def test_query_params(
     assert page.select_one('#gc-title').text.strip() == (
         heading
     )
+
+def test_should_render_welcome(client):
+    response = client.get(url_for('main.welcome'))
+    assert response.status_code == 200
+
+    page = BeautifulSoup(response.data.decode('utf-8'), 'html.parser')
+    assert page.h1.string == 'Welcome to GC Notify'
+
+    expected = "Create your first service"
+    link_text = page.find_all('a', {'class': 'button'})[0].text
+    assert link_text == expected
